refactor(daemon): fold prompt-templates into media/
Move prompt-templates.ts under the media/ folder so the prompt-template
reader lives with the other media modules it belongs with. It is a pure leaf
module (no relative imports), so the move is a byte-identical rename; only the
static-resource route importer and the colocated test are repointed. The
server.ts references are content-directory string literals, not module
imports, and are left untouched.

No public export surface change; no behavior change.
