+++ apps/daemon/src/media/prompt-templates.ts
